requirements.txt
Django==1.10.1
django-hitcount==1.2.2
django-extensions==1.7.3